pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zettelforge"
version = "2.1.1"
description = "ZettelForge: Agentic Memory System with vector search, knowledge graph, and synthesis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Patrick Roland", email = "patrick@groland.com"}
]
keywords = ["memory", "rag", "agent", "vector-search", "knowledge-graph", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies (Community edition — MIT licensed)
dependencies = [
    "lancedb>=0.5.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "tantivy>=0.11.0",
    "fastembed>=0.8.0",
    "llama-cpp-python>=0.3.0",
    "httpx>=0.25.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]

# Enterprise: install the separate zettelforge-enterprise package
# pip install zettelforge-enterprise
enterprise = [
    "zettelforge-enterprise>=2.1.0",
]

# Web UI (ThreatRecall)
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/rolandpg/zettelforge"
Documentation = "https://github.com/rolandpg/zettelforge/blob/main/docs/"
Repository = "https://github.com/rolandpg/zettelforge"
Issues = "https://github.com/rolandpg/zettelforge/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/zettelforge"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N"]
ignore = [
    "E501",   # line too long — handled by formatter
    "E402",   # module-level import not at top — needed for conditional imports
]

[tool.ruff.lint.per-file-ignores]
"src/zettelforge/__init__.py" = ["F401", "E402"]  # re-exports and conditional imports

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
